+++ CHANGELOG.md
@@ -15,6 +15,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Loading Scene System
 - Save / Load System
 
+## [0.1.1] - 2022-08-09
+### Changed
+- Enhancements to the `InteractorObject`.
+- The Input System (more specifically `KeyHoldInputController`) now tracks the order the keys were pressed, returning them in that order.
+- Code optimizations.
+- The `Pathfinder` has now an asynchronous method `FindShortestPathAsync()`.
+- The `KeyPathMovementInputListener` can now interrupt the movement, in order to start a new path movement.
+- The initial direction of a `WalkableObject` can now be specified in the inspector.
+
+### Fixed
+- Fixed missing documentation images.
+
 ## [0.1.0] - 2022-08-02
 ### Added
 - Initial Documentation: `README.md`, `CHANGELOG.md` and `Documentation~`.
@@ -40,4 +52,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Interaction and Trigger Systems
 - Sprite Animation
 
+[0.1.1]: https://github.com/skeltechnology/RPEST/releases/tag/v0.1.1
 [0.1.0]: https://github.com/skeltechnology/RPEST/releases/tag/v0.1.0